Consolidate common code in ContainerActivity
