+++ GreekReference/src/main/java/com/benlinskey/greekreference/AbstractDetailActivity.java
@@ -16,32 +16,22 @@
 
 package com.benlinskey.greekreference;
 
-import android.app.AlertDialog;
-import android.app.Dialog;
-import android.app.DialogFragment;
-import android.content.DialogInterface;
 import android.content.Intent;
-import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
 import android.support.v7.app.ActionBar;
-import android.support.v7.app.ActionBarActivity;
 import android.support.v7.widget.Toolbar;
-import android.text.Html;
-import android.text.method.LinkMovementMethod;
 import android.view.KeyEvent;
 import android.view.MenuItem;
-import android.widget.ScrollView;
-import android.widget.TextView;
+
+import com.benlinskey.greekreference.views.ContainerActivity;
 
 /**
  * The basic activity from which all detail activities inherit. This class 
  * contains a single {@link AbstractDetailFragment} and is only used on phones.
  */
-public abstract class AbstractDetailActivity extends ActionBarActivity {
-    // TODO: Some code is repeated from MainActivity here. It would be good to
-    // move this to a superclass or otherwise consolidate it somehow.
+public abstract class AbstractDetailActivity extends ContainerActivity {
 
     /** Stores the mode title. */
     protected CharSequence mTitle;
@@ -81,58 +71,6 @@ public abstract class AbstractDetailActivity extends ActionBarActivity {
             return true;
         }
         return super.onOptionsItemSelected(item);
-    }
-
-    /**
-     * Launches an email app that the user can use to send feedback about this app.
-     */
-    private void sendFeedback() {
-        Uri uri = Uri.fromParts("mailto", getString(R.string.feedback_email), null);
-        Intent intent = new Intent(Intent.ACTION_SENDTO, uri);
-        intent.putExtra(Intent.EXTRA_SUBJECT, getString(R.string.feedback_subject));
-        Intent chooser = Intent.createChooser(intent, getString(R.string.feedback_intent_chooser));
-        startActivity(chooser);
-    }
-
-    /**
-     * A {@link DialogFragment} containing help text.
-     */
-    public static class HelpDialogFragment extends DialogFragment {
-        
-        @Override
-        public Dialog onCreateDialog(Bundle savedInstanceState) {
-            AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
-            builder.setTitle(R.string.title_help);
-
-            TextView textView = new TextView(getActivity());
-            textView.setTextAppearance(getActivity(), android.R.style.TextAppearance_Medium);
-            textView.setTextColor(getResources().getColor(android.R.color.black));
-            textView.setPadding(25, 25, 25, 25);
-            
-            textView.setText(Html.fromHtml(getString(R.string.message_help)));
-            textView.setMovementMethod(LinkMovementMethod.getInstance());
-            
-            ScrollView scrollView = new ScrollView(getActivity());
-            scrollView.addView(textView);
-            builder.setView(scrollView);
-
-            builder.setPositiveButton(R.string.ok, new DialogInterface.OnClickListener() {
-                @Override
-                public void onClick(DialogInterface dialogInterface, int i) {
-                    dialogInterface.dismiss();
-                }
-            });
-            
-            return builder.create();
-        }
-    }
-
-    /**
-     * Displays a dialog fragment containing help text.
-     */
-    private void displayHelp() {
-        HelpDialogFragment dialogFragment = new HelpDialogFragment();
-        dialogFragment.show(getFragmentManager(), "help");
     }
 
     /**
