+++ GreekReference/src/main/java/com/benlinskey/greekreference/presenters/MainPresenter.java
@@ -206,10 +206,4 @@ public final class MainPresenter {
         mContext.startActivity(new Intent(mContext, SettingsActivity.class));
     }
 
-    public void onDisplayHelp() {
-        mView.displayHelp();
-    }
-
-
-
 }
